requirements.txt
flask==2.3.6
gunicorn==21.2.0
requests==2.34.0
python-dotenv==1.1.1
pandas==2.1.2
numpy==1.27.4
scipy==1.11.1
scikit-learn==1.3.2
